Refactor cost of carry
